Bug Fixing - Various Lighting, Shadows and Material Faces
Bug Fixing - Various Lighting, Shadows and Material Faces 26/03/22 03:35.

- Shadows are now more uniform as there were various broken artefacts.
- Lighting Probe was added to the entrance hallway as lighting beforehand didn't make sense (it was bright rather than matching the museum lobby).
- Various other minor bugs fixed.
- Chair added to desk. Due to crunch time, rather than it being an office-style chair, it is a duplicate of existing chairs in the scene. Box collider added to this.

NOTE: For more of a "uniform lighting" throughout the museum, various Probuilder objects had to be Merged. Therefore, as it stands, a few object names are not up to date. Due to crunch time, it is very unlikely that it will be updated before Deadline. However, this is noted as a learning experience and most likely updated after the deadline.